+++ src/macros.rs
@@ -7,6 +7,36 @@
 // licenses, and is:
 // Copyright Ⓒ 2015 The Rust Project Developers
 
+/// Log a message at the error level, flush the logger, and then use the same
+/// message to panic.
+///
+/// This will duplicate the message, once via the registered logger, then
+/// again via stderr for the panic (default handler). Since this is a fatal
+/// and presumably serious condition, potential duplication is of less concern
+/// than the risk of missing the message. This will always `panic!`, even if
+/// no logger is configured, or if error level messages aren't logged.
+///
+/// # Example
+///
+/// ```rust,should_panic
+/// # use std::time::{Duration, Instant};
+/// use tao_log::fatal;
+///
+/// # let td = Duration::new(0, 100_000_000);
+/// fatal!("shields compromised, core breach in {:?}!", td);
+/// // ^1 -- error level message: shields compromised, core breach in 100ms!
+/// // ^2 -- panic:               shields compromised, core breach in 100ms!
+/// ```
+#[macro_export]
+macro_rules! fatal {
+    (target: $target:expr, $($arg:tt)+) => (
+        $crate::__tao_fatal!($target, $($arg)+)
+    );
+    ($($arg:tt)+) => (
+        $crate::__tao_fatal!(module_path!(), $($arg)+)
+    );
+}
+
 /// Log an expression and its value at any specified level.
 ///
 /// Logs with the optional or default (module path of use) target, specified
@@ -77,6 +107,21 @@ macro_rules! debugv {
 #[macro_export]
 macro_rules! tracev {
     ($($arg:tt)+) => ($crate::__tao_logv!($crate::log::Level::Trace, $($arg)+))
+}
+
+// Helper macro for `fatal!`
+#[doc(hidden)]
+#[macro_export]
+macro_rules! __tao_fatal {
+    ($target:expr, $($arg:tt)+) => (
+        match format_args!($($arg)+) {
+            args => {
+                $crate::error!(target: $target, "{}", args);
+                $crate::log::logger().flush();
+                panic!("{}", args);
+            }
+        }
+    );
 }
 
 // Helper macro for the -v macros, handling the permutations of optional